Make work with V3 Lynxmotion PS2
The new version of the Lynxmotion PS2 does not work well with current
PS2 reading code that we had for the Basic Atom Pros.  In particular it
did not like how we read the current mode and then started up a new read
to read everything.

Changed to instead look for a timeout by seeing if it had been 1.5
seconds since our last read (like Arduino PS2 library).  Also later
verify that first byte we read was
an 0x5A...  Not perfect but got it reading.